+++ @utahdts/utah-design-system/react/components/forms/useFormContext.js
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import FormContext from './FormContext';
+
+export default function useFormContext() {
+  return useContext(FormContext);
+}