+++ @utahdts/utah-design-system/react/components/table/TableCell.jsx
@@ -14,16 +14,12 @@ const propTypes = {
   className: PropTypes.string,
   id: PropTypes.string,
   innerRef: RefShape,
-  onClick: PropTypes.func,
-  onDoubleClick: PropTypes.func,
 };
 const defaultProps = {
   children: null,
   className: null,
   id: null,
   innerRef: null,
-  onClick: null,
-  onDoubleClick: null,
 };
 
 /**
@@ -33,8 +29,6 @@ const defaultProps = {
  * @param {string | null} [props.className]
  * @param {string | null} [props.id]
  * @param {React.RefObject | null} [props.innerRef]
- * @param {((param: {e: Event}) => void) | null} [props.onClick]
- * @param {((param: {e: Event}) => void) | null} [props.onDoubleClick]
  * @returns {JSX.Element}
  */
 function TableCell({
@@ -42,8 +36,6 @@ function TableCell({
   className = null,
   id = null,
   innerRef = null,
-  onClick = null,
-  onDoubleClick = null,
   ...rest
 }) {
   return (
@@ -52,10 +44,6 @@ function TableCell({
       className={joinClassNames('table__cell', className)}
       id={id ?? undefined}
       ref={innerRef}
-      // @ts-ignore
-      onClick={onClick ?? undefined}
-      // @ts-ignore
-      onDoubleClick={onDoubleClick ?? undefined}
       {...rest}
     >
       {children}