+++ utah-design-system-website/src/react/components/websiteContent/library/components/forms/CharacterCount/CharacterCountDocumentation.jsx
@@ -115,7 +115,7 @@ function CharacterCountDocumentation() {
 
       <h2 id="section-settings-props">Settings and Props</h2>
       <div className="documentation-content--small-text">
-        <TabGroup defaultValue="button-props-react">
+        <TabGroup defaultValue="button-props-css">
           <TabList>
             <Tab id="button-props-css">CSS</Tab>
             <Tab id="button-props-react">React</Tab>