+++ utah-design-system-website/src/react/components/websiteContent/library/components/forms/Select/SelectCssClassesDocumentation.jsx
@@ -1,0 +1,31 @@
+import React from 'react';
+import {
+  Table, TableBody, TableCell, TableHead, TableHeadCell, TableHeadRow, TableRow, TableWrapper
+} from '@utahdts/utah-design-system';
+
+/** @returns {JSX.Element} */
+function SelectCssClassesDocumentation() {
+  return (
+    <TableWrapper>
+      <Table className="table--lines-x table--full-width">
+        <TableHead>
+          <TableHeadRow>
+            <TableHeadCell className="text-left css-classes">CSS Classes</TableHeadCell>
+            <TableHeadCell className="text-left ">Description</TableHeadCell>
+          </TableHeadRow>
+        </TableHead>
+        <TableBody>
+          <TableRow>
+            <TableCell><code>.select-input--clear-icon-visible</code></TableCell>
+            <TableCell>
+              This class is used conjointly with the <code>isClearable</code> property.
+              It adds extra padding to the right side to accommodate the clear icon.
+            </TableCell>
+          </TableRow>
+        </TableBody>
+      </Table>
+    </TableWrapper>
+  );
+}
+
+export default SelectCssClassesDocumentation;