+++ utah-design-system-website/src/react/components/websiteContent/library/components/forms/TextArea/TextAreaCssClassesDocumentation.jsx
@@ -17,7 +17,10 @@ function TextAreaCssClassesDocumentation() {
         <TableBody>
           <TableRow>
             <TableCell><code>.text-area--clear-icon-visible</code></TableCell>
-            <TableCell>This class is used conjointly with the <code>isClearable</code> property.</TableCell>
+            <TableCell>
+              This class is used conjointly with the <code>isClearable</code> property.
+              It adds extra padding to the right side to accommodate the clear icon.
+            </TableCell>
           </TableRow>
         </TableBody>
       </Table>